Extract MultiParentCasperCommunicationSpec
